Add username to chat messages and events
In order to add some user identification to chat messages and events, the code was changed to include `username` along with `message`. Now, `username` is passed to the `NewMessage` class and to `DomElements` methods, which enables chat messages to display user's name and user's list updates when a user arrives or exits can refer back to the specific user.